Refactor opencv display window code into utils/display.py
