+++ utils/display.py
@@ -1,0 +1,32 @@
+
+"""display.py
+"""
+
+
+import cv2
+
+
+def open_window(window_name, width, height, title):
+    """Open the display window."""
+    cv2.namedWindow(window_name, cv2.WINDOW_NORMAL)
+    cv2.resizeWindow(window_name, width, height)
+    cv2.moveWindow(window_name, 0, 0)
+    cv2.setWindowTitle(window_name, title)
+
+
+def show_help_text(img, help_text):
+    """Draw help text on image."""
+    cv2.putText(img, help_text, (11, 20), cv2.FONT_HERSHEY_PLAIN, 1.0,
+                (32, 32, 32), 4, cv2.LINE_AA)
+    cv2.putText(img, help_text, (10, 20), cv2.FONT_HERSHEY_PLAIN, 1.0,
+                (240, 240, 240), 1, cv2.LINE_AA)
+
+
+def set_display(window_name, full_scrn):
+    """Set disply window to either full screen or normal."""
+    if full_scrn:
+        cv2.setWindowProperty(window_name, cv2.WND_PROP_FULLSCREEN,
+                              cv2.WINDOW_FULLSCREEN)
+    else:
+        cv2.setWindowProperty(window_name, cv2.WND_PROP_FULLSCREEN,
+                              cv2.WINDOW_NORMAL)
